pyproject.toml
[tool.poetry]
name = "Reina"
version = "0.1.0"
description = "A fork of Beryl built for production and enterprise use"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiodns = "^3.0.0"
python-dotenv = "^0.21.0"
uvloop = "^0.16.0"
requests = "^2.28.1"
pysimdjson = "^5.0.2"
ffmpeg-python = "^0.2.0"
orjson = "^3.8.1"
python-dateutil = "^2.8.2"
rin-exceptions = "^1.0.2"
Brotli = "^1.0.9"
numpy = "^1.23.4"
asyncpraw = "^7.6.0"
tortoise-orm = {extras = ["accel"], version = "^0.19.2"}
aiocron = "^1.8"
gql = {extras = ["aiohttp"], version = "^3.4.0"}
py-cord = {extras = ["voice"], version = "^2.2.2"}
charset-normalizer = "^2.1.1"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
pyinstrument = "^4.4.0"

[tool.isort]
profile = 'black'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
